requirements.txt
flask
librosa
pandas
numpy
tensorflow-cpu
joblib
scipy
soundfile==0.12.1
# setuptools==67.7.2